RDKCOM-378: set stdout to unbuffered mode
Set stdout to unbuffered mode as the dmcli app needs to generate
terminal output which is not always line based (e.g. the "dmcli>"
prompt) and therefore does not always work as expected if stdout is
left in the default line buffered mode.
Signed-off-by: Andre McCurdy <armccurdy@gmail.com>

Reason for change: set stdout to unbuffered mode
Test Procedure: None.
Risks: Low.

Change-Id: I8dcce4f053fa150ae32747df2a2b71230c578248